fix: stale comment + real(8)→real(real64) in irrigation adapter
- irrigation_config: comment now references read_csv_table (the
  removed read_csv_date_reals was a stale post-rename leftover).
- config_to_variables irrigation block: use iso_fortran_env real64
  to match the BBC blocks; eliminates the kind-literal inconsistency.

+++ src/io/toml/config_to_variables.f90
@@ -841,7 +841,8 @@ contains
             block
                use csv_reader_mod, only: read_csv_table
                use error_mod, only: error_collection_t
-               real(8), allocatable :: csv_table(:,:)
+               use iso_fortran_env, only: real64
+               real(real64), allocatable :: csv_table(:,:)
                type(error_collection_t) :: csv_errs
                integer :: k_csv, nrows_csv
                character(len=5) :: irrig_header(4)
